+++ src/racket/gc2/gc2_obj.h
@@ -9,16 +9,16 @@
 # define LOG_APAGE_SIZE 14
 #endif
 typedef struct objhead {
-  unsigned long hash      : ((8 * sizeof(unsigned long)) - (4+3+LOG_APAGE_SIZE) );
+  intptr_t hash      : ((8 * sizeof(intptr_t)) - (4+3+LOG_APAGE_SIZE) );
   /* the type and size of the object */
-  unsigned long type      : 3;
+  intptr_t type      : 3;
   /* these are the various mark bits we use */
-  unsigned long mark      : 1;
-  unsigned long btc_mark  : 1;
+  intptr_t mark      : 1;
+  intptr_t btc_mark  : 1;
   /* these are used for compaction et al*/
-  unsigned long moved     : 1;
-  unsigned long dead      : 1;
-  unsigned long size      : LOG_APAGE_SIZE;
+  intptr_t moved     : 1;
+  intptr_t dead      : 1;
+  intptr_t size      : LOG_APAGE_SIZE;
 } objhead;
 
 #define OBJHEAD_SIZE (sizeof(objhead))
